test: move GFM tests to gfm-markdown.test.ts
